Created the invokeService for invoking chaincode.  Added config
